Move utilities into a package
Define a package for utilities. Rename the module
defining utilities for url encoding to
datalad_registry/utils/url_encoder.py
