+++ datalad_registry/dataset_urls.py
@@ -11,7 +11,7 @@ from flask import Blueprint, jsonify, request
 
 from datalad_registry import tasks
 from datalad_registry.models import URL, db
-from datalad_registry.utils import InvalidURL, url_decode
+from datalad_registry.utils.url_encoder import InvalidURL, url_decode
 
 lgr = logging.getLogger(__name__)
 bp = Blueprint("dataset_urls", __name__, url_prefix="/v1/datasets/")
